+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/command/application/controller/RecruitmentBoardReplyController.java
@@ -1,0 +1,4 @@
+package com.springcooler.sgma.recruitmentboardreply.command.application.controller;
+
+public class RecruitmentBoardReply {
+}
