+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/command/application/dto/RecruitmentBoardReplyDTO.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.command.application.dto;public class RecruitmentBoardReplyDTO {
+}
