+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/command/application/service/RecruitmentBoardReplyService.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.command.application;public interface RecruitmentBoardReplyService {
+}
