+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/command/domain/aggregate/ActiveStatus.java
@@ -1,0 +1,5 @@
+package com.springcooler.sgma.recruitmentboardreply.command.domain.aggregate;
+
+public enum ActiveStatus1 {
+    ACTIVE,INACTIVE
+}
