+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/command/domain/aggregate/AnonymousStatus.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.command.domain.aggregate;public enum AnonymousStatus1 {
+}
