+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/command/domain/repository/RecruitmentBoardReplyRepository.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.command.domain.repository;public class RecruitmentBoardReplyRepository {
+}
