+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/query/controller/RecruitmentBoardReplyController.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.query.controller;public class RecruitmentController {
+}
