+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/query/dto/RecruitmentBoardReplyDTO.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.query.dto;public class RecruitmentBoardReplyDTO {
+}
