+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/query/repository/RecruitmentBoardReplyMapper.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.query.repository;public class RecruitmentBoardReplyMapper {
+}
