+++ src/main/java/com/springcooler/sgma/recruitmentboardreply/query/service/RecruitmentBoardReplyService.java
@@ -1,0 +1,2 @@
+package com.springcooler.sgma.recruitmentboardreply.query.service;public class RecruitmentBoardReplyService {
+}
